Split workspace project parity gates
